// src/codegen/MsgGroupUpdate.ts
/* tslint:disable */
/* eslint-disable */
/*
 * Autogenerated by @creditkarma/thrift-typescript v3.7.6
 * DO NOT EDIT UNLESS YOU ARE SURE THAT YOU KNOW WHAT YOU ARE DOING
*/
import * as thrift from "thrift";
import * as PayloadGroup from "./PayloadGroup";
export interface IMsgGroupUpdateArgs {
    name?: string;
    group?: PayloadGroup.PayloadGroup;
}
export class MsgGroupUpdate {
    public name?: string;
    public group?: PayloadGroup.PayloadGroup;
    constructor(args?: IMsgGroupUpdateArgs) {
        if (args != null && args.name != null) {
            this.name = args.name;
        }
        if (args != null && args.group != null) {
            this.group = args.group;
        }
    }
    public write(output: thrift.TProtocol): void {
        output.writeStructBegin("MsgGroupUpdate");
        if (this.name != null) {
            output.writeFieldBegin("name", thrift.Thrift.Type.STRING, 1);
            output.writeString(this.name);
            output.writeFieldEnd();
        }
        if (this.group != null) {
            output.writeFieldBegin("group", thrift.Thrift.Type.STRUCT, 2);
            this.group.write(output);
            output.writeFieldEnd();
        }
        output.writeFieldStop();
        output.writeStructEnd();
        return;
    }
    public static read(input: thrift.TProtocol): MsgGroupUpdate {
        input.readStructBegin();
        let _args: any = {};
        while (true) {
            const ret: thrift.TField = input.readFieldBegin();
            const fieldType: thrift.Thrift.Type = ret.ftype;
            const fieldId: number = ret.fid;
            if (fieldType === thrift.Thrift.Type.STOP) {
                break;
            }
            switch (fieldId) {
                case 1:
                    if (fieldType === thrift.Thrift.Type.STRING) {
                        const value_1: string = input.readString();
                        _args.name = value_1;
                    }
                    else {
                        input.skip(fieldType);
                    }
                    break;
                case 2:
                    if (fieldType === thrift.Thrift.Type.STRUCT) {
                        const value_2: PayloadGroup.PayloadGroup = PayloadGroup.PayloadGroup.read(input);
                        _args.group = value_2;
                    }
                    else {
                        input.skip(fieldType);
                    }
                    break;
                default: {
                    input.skip(fieldType);
                }
            }
            input.readFieldEnd();
        }
        input.readStructEnd();
        return new MsgGroupUpdate(_args);
    }
}