// src/codegen/RspPlayerLogin.ts
/* tslint:disable */
/* eslint-disable */
/*
 * Autogenerated by @creditkarma/thrift-typescript v3.7.6
 * DO NOT EDIT UNLESS YOU ARE SURE THAT YOU KNOW WHAT YOU ARE DOING
*/
import * as thrift from "thrift";
import * as Error from "./Error";
import * as PayloadPlayer from "./PayloadPlayer";
export interface IRspPlayerLoginArgs {
    result?: Error.Error;
    player?: PayloadPlayer.PayloadPlayer;
}
export class RspPlayerLogin {
    public result?: Error.Error;
    public player?: PayloadPlayer.PayloadPlayer;
    constructor(args?: IRspPlayerLoginArgs) {
        if (args != null && args.result != null) {
            this.result = args.result;
        }
        if (args != null && args.player != null) {
            this.player = args.player;
        }
    }
    public write(output: thrift.TProtocol): void {
        output.writeStructBegin("RspPlayerLogin");
        if (this.result != null) {
            output.writeFieldBegin("result", thrift.Thrift.Type.I32, 1);
            output.writeI32(this.result);
            output.writeFieldEnd();
        }
        if (this.player != null) {
            output.writeFieldBegin("player", thrift.Thrift.Type.STRUCT, 6);
            this.player.write(output);
            output.writeFieldEnd();
        }
        output.writeFieldStop();
        output.writeStructEnd();
        return;
    }
    public static read(input: thrift.TProtocol): RspPlayerLogin {
        input.readStructBegin();
        let _args: any = {};
        while (true) {
            const ret: thrift.TField = input.readFieldBegin();
            const fieldType: thrift.Thrift.Type = ret.ftype;
            const fieldId: number = ret.fid;
            if (fieldType === thrift.Thrift.Type.STOP) {
                break;
            }
            switch (fieldId) {
                case 1:
                    if (fieldType === thrift.Thrift.Type.I32) {
                        const value_1: Error.Error = input.readI32();
                        _args.result = value_1;
                    }
                    else {
                        input.skip(fieldType);
                    }
                    break;
                case 6:
                    if (fieldType === thrift.Thrift.Type.STRUCT) {
                        const value_2: PayloadPlayer.PayloadPlayer = PayloadPlayer.PayloadPlayer.read(input);
                        _args.player = value_2;
                    }
                    else {
                        input.skip(fieldType);
                    }
                    break;
                default: {
                    input.skip(fieldType);
                }
            }
            input.readFieldEnd();
        }
        input.readStructEnd();
        return new RspPlayerLogin(_args);
    }
}