// src/codegen/CommandType.ts
/* tslint:disable */
/* eslint-disable */
/*
 * Autogenerated by @creditkarma/thrift-typescript v3.7.6
 * DO NOT EDIT UNLESS YOU ARE SURE THAT YOU KNOW WHAT YOU ARE DOING
*/
export enum CommandType {
    RSP_PLAYER_LOGIN = 102,
    REQ_PLAYER_REQUEST = 109,
    RSP_PLAYER_REQUEST = 110,
    RSP_GROUP_LISTENER_REGISTER = 314,
    MSG_GROUP_UPDATE = 350
}